+++ sdk/python/feast/infra/gcp.py
@@ -6,8 +6,12 @@ from typing import Any, Callable, Dict, Iterator, List, Optional, Sequence, Tupl
 import mmh3
 import pandas
 import pyarrow
+from google.auth.exceptions import DefaultCredentialsError
+from tqdm import tqdm
 
 from feast import FeatureTable, utils
+from feast.entity import Entity
+from feast.errors import FeastProviderLoginError
 from feast.feature_view import FeatureView
 from feast.infra.key_encoding_utils import serialize_entity_key
 from feast.infra.offline_stores.helpers import get_offline_store_from_sources
@@ -37,16 +41,24 @@ class GcpProvider(Provider):
     def _initialize_client(self):
         from google.cloud import datastore
 
-        if self._gcp_project_id is not None:
-            return datastore.Client(self._gcp_project_id)
-        else:
-            return datastore.Client()
+        try:
+            if self._gcp_project_id is not None:
+                return datastore.Client(self._gcp_project_id)
+            else:
+                return datastore.Client()
+        except DefaultCredentialsError as e:
+            raise FeastProviderLoginError(
+                str(e)
+                + '\nIt may be necessary to run "gcloud auth application-default login" if you would like to use your local Google Cloud account'
+            )
 
     def update_infra(
         self,
         project: str,
         tables_to_delete: Sequence[Union[FeatureTable, FeatureView]],
         tables_to_keep: Sequence[Union[FeatureTable, FeatureView]],
+        entities_to_delete: Sequence[Entity],
+        entities_to_keep: Sequence[Entity],
         partial: bool,
     ):
         from google.cloud import datastore
@@ -69,7 +81,10 @@ class GcpProvider(Provider):
             client.delete(key)
 
     def teardown_infra(
-        self, project: str, tables: Sequence[Union[FeatureTable, FeatureView]]
+        self,
+        project: str,
+        tables: Sequence[Union[FeatureTable, FeatureView]],
+        entities: Sequence[Entity],
     ) -> None:
         client = self._initialize_client()
 
@@ -104,6 +119,7 @@ class GcpProvider(Provider):
         project: str,
         table: Union[FeatureTable, FeatureView],
         entity_keys: List[EntityKeyProto],
+        requested_features: List[str] = None,
     ) -> List[Tuple[Optional[datetime], Optional[Dict[str, ValueProto]]]]:
         client = self._initialize_client()
 
@@ -132,6 +148,7 @@ class GcpProvider(Provider):
         end_date: datetime,
         registry: Registry,
         project: str,
+        tqdm_builder: Callable[[int], tqdm],
     ) -> None:
         entities = []
         for entity_name in feature_view.entities:
@@ -164,7 +181,10 @@ class GcpProvider(Provider):
         join_keys = [entity.join_key for entity in entities]
         rows_to_write = _convert_arrow_to_proto(table, feature_view, join_keys)
 
-        self.online_write_batch(project, feature_view, rows_to_write, None)
+        with tqdm_builder(len(rows_to_write)) as pbar:
+            self.online_write_batch(
+                project, feature_view, rows_to_write, lambda x: pbar.update(x)
+            )
 
         feature_view.materialization_intervals.append((start_date, end_date))
         registry.apply_feature_view(feature_view, project)