+++ sdk/python/tests/test_offline_online_store_consistency.py
@@ -4,7 +4,7 @@ import time
 import uuid
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Iterator, Tuple, Union
+from typing import Iterator, Optional, Tuple, Union
 
 import pandas as pd
 import pytest
@@ -26,7 +26,7 @@ def create_dataset() -> pd.DataFrame:
     ts = pd.Timestamp(now).round("ms")
     data = {
         "id": [1, 2, 1, 3, 3],
-        "value": [0.1, 0.2, 0.3, 4, 5],
+        "value": [0.1, None, 0.3, 4, 5],
         "ts_1": [
             ts - timedelta(hours=4),
             ts,
@@ -141,19 +141,56 @@ def prep_local_fs_and_fv() -> Iterator[Tuple[FeatureStore, FeatureView]]:
             yield fs, fv
 
 
+@contextlib.contextmanager
+def prep_redis_fs_and_fv() -> Iterator[Tuple[FeatureStore, FeatureView]]:
+    with tempfile.NamedTemporaryFile(suffix=".parquet") as f:
+        df = create_dataset()
+        f.close()
+        df.to_parquet(f.name)
+        file_source = FileSource(
+            file_format=ParquetFormat(),
+            file_url=f"file://{f.name}",
+            event_timestamp_column="ts",
+            created_timestamp_column="created_ts",
+            date_partition_column="",
+            field_mapping={"ts_1": "ts", "id": "driver_id"},
+        )
+        fv = get_feature_view(file_source)
+        e = Entity(
+            name="driver",
+            description="id for driver",
+            join_key="driver_id",
+            value_type=ValueType.INT32,
+        )
+        with tempfile.TemporaryDirectory() as repo_dir_name, tempfile.TemporaryDirectory():
+            config = RepoConfig(
+                registry=str(Path(repo_dir_name) / "registry.db"),
+                project=f"test_bq_correctness_{str(uuid.uuid4()).replace('-', '')}",
+                provider="redis",
+            )
+            fs = FeatureStore(config=config)
+            fs.apply([fv, e])
+
+            yield fs, fv
+
+
 # Checks that both offline & online store values are as expected
 def check_offline_and_online_features(
     fs: FeatureStore,
     fv: FeatureView,
     driver_id: int,
     event_timestamp: datetime,
-    expected_value: float,
+    expected_value: Optional[float],
 ) -> None:
     # Check online store
     response_dict = fs.get_online_features(
         [f"{fv.name}:value"], [{"driver": driver_id}]
     ).to_dict()
-    assert abs(response_dict[f"{fv.name}__value"][0] - expected_value) < 1e-6
+
+    if expected_value:
+        assert abs(response_dict[f"{fv.name}__value"][0] - expected_value) < 1e-6
+    else:
+        assert response_dict[f"{fv.name}__value"][0] is None
 
     # Check offline store
     df = fs.get_historical_features(
@@ -163,7 +200,11 @@ def check_offline_and_online_features(
         feature_refs=[f"{fv.name}:value"],
     ).to_df()
 
-    assert abs(df.to_dict()[f"{fv.name}__value"][0] - expected_value) < 1e-6
+    if expected_value:
+        assert abs(df.to_dict()[f"{fv.name}__value"][0] - expected_value) < 1e-6
+    else:
+        df = df.where(pd.notnull(df), None)
+        assert df.to_dict()[f"{fv.name}__value"][0] is None
 
 
 def run_offline_online_store_consistency_test(
@@ -179,6 +220,10 @@ def run_offline_online_store_consistency_test(
     # check result of materialize()
     check_offline_and_online_features(
         fs=fs, fv=fv, driver_id=1, event_timestamp=end_date, expected_value=0.3
+    )
+
+    check_offline_and_online_features(
+        fs=fs, fv=fv, driver_id=2, event_timestamp=end_date, expected_value=None
     )
 
     # check prior value for materialize_incremental()
@@ -201,6 +246,12 @@ def run_offline_online_store_consistency_test(
 )
 def test_bq_offline_online_store_consistency(bq_source_type: str):
     with prep_bq_fs_and_fv(bq_source_type) as (fs, fv):
+        run_offline_online_store_consistency_test(fs, fv)
+
+
+@pytest.mark.integration
+def test_redis_offline_online_store_consistency():
+    with prep_redis_fs_and_fv() as (fs, fv):
         run_offline_online_store_consistency_test(fs, fv)