+++ sdk/python/tests/test_online_retrieval.py
@@ -5,6 +5,7 @@ from datetime import datetime
 import pytest
 
 from feast import FeatureStore, RepoConfig
+from feast.errors import FeatureViewNotFoundException
 from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.repo_config import RegistryConfig
@@ -31,7 +32,7 @@ def test_online() -> None:
             join_keys=["driver"], entity_values=[ValueProto(int64_val=1)]
         )
         provider.online_write_batch(
-            project=store.config.project,
+            project=store.project,
             table=driver_locations_fv,
             data=[
                 (
@@ -51,7 +52,7 @@ def test_online() -> None:
             join_keys=["customer"], entity_values=[ValueProto(int64_val=5)]
         )
         provider.online_write_batch(
-            project=store.config.project,
+            project=store.project,
             table=customer_profile_fv,
             data=[
                 (
@@ -73,7 +74,7 @@ def test_online() -> None:
             entity_values=[ValueProto(int64_val=5), ValueProto(int64_val=1)],
         )
         provider.online_write_batch(
-            project=store.config.project,
+            project=store.project,
             table=customer_driver_combined_fv,
             data=[
                 (
@@ -116,7 +117,7 @@ def test_online() -> None:
         assert "customer_driver_combined__trips" in result
 
         # invalid table reference
-        with pytest.raises(ValueError):
+        with pytest.raises(FeatureViewNotFoundException):
             store.get_online_features(
                 feature_refs=["driver_locations_bad:lon"], entity_rows=[{"driver": 1}],
             )
@@ -129,7 +130,7 @@ def test_online() -> None:
                     path=store.config.registry, cache_ttl_seconds=cache_ttl
                 ),
                 online_store=store.config.online_store,
-                project=store.config.project,
+                project=store.project,
                 provider=store.config.provider,
             )
         )
@@ -188,7 +189,7 @@ def test_online() -> None:
                     path=store.config.registry, cache_ttl_seconds=0
                 ),
                 online_store=store.config.online_store,
-                project=store.config.project,
+                project=store.project,
                 provider=store.config.provider,
             )
         )