+++ CHANGELOG.md
@@ -1,8 +1,59 @@
 # Changelog
 
-## [0.10.0](https://github.com/feast-dev/feast/tree/0.10.0) (2021-04-15)
+## [v0.10.3](https://github.com/feast-dev/feast/tree/v0.10.3) (2021-04-21)
 
-[Full Changelog](https://github.com/feast-dev/feast/compare/v0.9.5...0.10.0)
+[Full Changelog](https://github.com/feast-dev/feast/compare/v0.10.2...v0.10.3)
+
+**Implemented enhancements:**
+
+- Add support for third party providers [\#1501](https://github.com/feast-dev/feast/pull/1501) ([tsotnet](https://github.com/tsotnet))
+- Infer entity dataframe event timestamp column [\#1495](https://github.com/feast-dev/feast/pull/1495) ([jklegar](https://github.com/jklegar))
+- Allow Feast apply to import files recursively \(and add .feastignore\) [\#1482](https://github.com/feast-dev/feast/pull/1482) ([tsotnet](https://github.com/tsotnet))
+
+**Fixed bugs:**
+
+- Bump supported Python version to 3.7 [\#1504](https://github.com/feast-dev/feast/pull/1504) ([tsotnet](https://github.com/tsotnet))
+- Fix bug in allowing empty repositories to be applied to a GCS registry [\#1488](https://github.com/feast-dev/feast/pull/1488) ([woop](https://github.com/woop))
+
+**Merged pull requests:**
+
+- Add a fixed timestamp to quickstart data [\#1513](https://github.com/feast-dev/feast/pull/1513) ([jklegar](https://github.com/jklegar))
+- Make gcp imports optional [\#1512](https://github.com/feast-dev/feast/pull/1512) ([jklegar](https://github.com/jklegar))
+- Fix documentation inconsistency [\#1510](https://github.com/feast-dev/feast/pull/1510) ([jongillham](https://github.com/jongillham))
+- pre-commit command typo fix in CONTRIBUTING.md [\#1506](https://github.com/feast-dev/feast/pull/1506) ([mavysavydav](https://github.com/mavysavydav))
+- Add optional telemetry to other CLI commands [\#1505](https://github.com/feast-dev/feast/pull/1505) ([jklegar](https://github.com/jklegar))
+- Pass entities information to Provider [\#1498](https://github.com/feast-dev/feast/pull/1498) ([MattDelac](https://github.com/MattDelac))
+- Update broken urls in contributing.md [\#1489](https://github.com/feast-dev/feast/pull/1489) ([tedhtchang](https://github.com/tedhtchang))
+- Python docs formatting fixes [\#1473](https://github.com/feast-dev/feast/pull/1473) ([jklegar](https://github.com/jklegar))
+
+## [v0.10.2](https://github.com/feast-dev/feast/tree/v0.10.2) (2021-04-21)
+
+[Full Changelog](https://github.com/feast-dev/feast/compare/v0.10.1...v0.10.2)
+
+**Fixed bugs:**
+
+- Fix bug in allowing empty repositories to be applied to a GCS registry [\#1488](https://github.com/feast-dev/feast/pull/1488) ([woop](https://github.com/woop))
+
+## [v0.10.1](https://github.com/feast-dev/feast/tree/v0.10.1) (2021-04-21)
+
+[Full Changelog](https://github.com/feast-dev/feast/compare/v0.10.0...v0.10.1)
+
+**Fixed bugs:**
+
+- Fix time zone issue with get\_historical\_features [\#1475](https://github.com/feast-dev/feast/pull/1475) ([tsotnet](https://github.com/tsotnet))
+
+**Merged pull requests:**
+
+- Improve exception handling, logging, and validation [\#1477](https://github.com/feast-dev/feast/pull/1477) ([woop](https://github.com/woop))
+- Remove duped pic [\#1476](https://github.com/feast-dev/feast/pull/1476) ([YikSanChan](https://github.com/YikSanChan))
+- Fix created timestamp related errors for BigQuery source [\#1474](https://github.com/feast-dev/feast/pull/1474) ([jklegar](https://github.com/jklegar))
+- Remove unnecessary MAVEN\_CONFIG [\#1472](https://github.com/feast-dev/feast/pull/1472) ([danielsiwiec](https://github.com/danielsiwiec))
+- Fix CLI entities command & add feature-views command [\#1471](https://github.com/feast-dev/feast/pull/1471) ([tsotnet](https://github.com/tsotnet))
+
+
+## [v0.10.0](https://github.com/feast-dev/feast/tree/0.10.0) (2021-04-15)
+
+[Full Changelog](https://github.com/feast-dev/feast/compare/v0.9.5...v0.10.0)
 
 ** Implemented enhancements:**